+++ lib/WorseReflection/Core/Inference/FunctionStub.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Inference;
+
+use Microsoft\PhpParser\Node\DelimitedList\ArgumentExpressionList;
+
+interface FunctionStub
+{
+    public function resolve(NodeContextResolver $resolver, Frame $frame, NodeContext $context, ArgumentExpressionList $node): NodeContext;
+}