+++ lib/WorseReflection/Core/Inference/Resolver/ParameterResolver.php
@@ -3,12 +3,9 @@
 namespace Phpactor\WorseReflection\Core\Inference\Resolver;
 
 use Microsoft\PhpParser\Node;
-use Microsoft\PhpParser\Node\DelimitedList\QualifiedNameList;
 use Microsoft\PhpParser\Node\Expression\AnonymousFunctionCreationExpression;
 use Microsoft\PhpParser\Node\MethodDeclaration;
 use Microsoft\PhpParser\Node\Parameter;
-use Microsoft\PhpParser\Node\QualifiedName;
-use Microsoft\PhpParser\Token;
 use Phpactor\WorseReflection\Core\Exception\CouldNotResolveNode;
 use Phpactor\WorseReflection\Core\Exception\ItemNotFound;
 use Phpactor\WorseReflection\Core\Inference\Frame;
@@ -19,7 +16,6 @@ use Phpactor\WorseReflection\Core\Inference\Symbol;
 use Phpactor\WorseReflection\Core\Inference\NodeContextResolver;
 use Phpactor\WorseReflection\Core\TypeFactory;
 use Phpactor\WorseReflection\Core\Util\NodeUtil;
-use Phpactor\WorseReflection\Core\Util\QualifiedNameListUtil;
 use Phpactor\WorseReflection\Reflector;
 
 class ParameterResolver implements Resolver
@@ -36,21 +32,8 @@ class ParameterResolver implements Resolver
         }
 
         $typeDeclaration = $node->typeDeclarationList;
-        $type = TypeFactory::unknown();
-        if ($typeDeclaration instanceof QualifiedNameList) {
-            $typeDeclaration = QualifiedNameListUtil::firstQualifiedNameOrToken($typeDeclaration);
-        }
 
-        if ($typeDeclaration instanceof QualifiedName) {
-            $type = $resolver->resolveNode($frame, $typeDeclaration)->type();
-        }
-
-        if ($typeDeclaration instanceof Token) {
-            $type = TypeFactory::fromStringWithReflector(
-                (string)$typeDeclaration->getText($node->getFileContents()),
-                $resolver->reflector(),
-            );
-        }
+        $type = NodeUtil::typeFromQualfiedNameLike($resolver->reflector(), $node, $node->typeDeclarationList);
 
         if ($node->questionToken) {
             $type = TypeFactory::nullable($type);