+++ lib/WorseReflection/Core/Inference/Resolver/QualifiedNameResolver.php
@@ -7,6 +7,7 @@ use Microsoft\PhpParser\Node\Expression\CallExpression;
 use Microsoft\PhpParser\Node\QualifiedName;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
 use Phpactor\WorseReflection\Core\Inference\Frame;
+use Phpactor\WorseReflection\Core\Inference\FunctionStubRegistry;
 use Phpactor\WorseReflection\Core\Inference\NodeToTypeConverter;
 use Phpactor\WorseReflection\Core\Inference\NodeContext;
 use Phpactor\WorseReflection\Core\Inference\NodeContextFactory;
@@ -24,19 +25,25 @@ class QualifiedNameResolver implements Resolver
 
     private NodeToTypeConverter $nodeTypeConverter;
 
+    private FunctionStubRegistry $registry;
 
-    public function __construct(FunctionReflector $reflector, NodeToTypeConverter $nodeTypeConverter)
-    {
+    public function __construct(
+        FunctionReflector $reflector,
+        FunctionStubRegistry $registry,
+        NodeToTypeConverter $nodeTypeConverter
+    ) {
         $this->reflector = $reflector;
         $this->nodeTypeConverter = $nodeTypeConverter;
+        $this->registry = $registry;
     }
 
     public function resolve(NodeContextResolver $resolver, Frame $frame, Node $node): NodeContext
     {
         assert($node instanceof QualifiedName);
 
-        if ($node->parent instanceof CallExpression) {
-            $name = $node->getResolvedName() ?: $node;
+        $parent = $node->parent;
+        if ($parent instanceof CallExpression) {
+            $name = $node->getNamespacedName();
             $name = Name::fromString((string) $name);
             $context = NodeContextFactory::create(
                 $name->short(),
@@ -46,6 +53,13 @@ class QualifiedNameResolver implements Resolver
                     'symbol_type' => Symbol::FUNCTION,
                 ]
             );
+
+            $stub = $this->registry->get($name);
+
+            $arguments = $parent->argumentExpressionList;
+            if ($stub && $arguments) {
+                return $stub->resolve($resolver, $frame, $context, $arguments);
+            }
 
             try {
                 $function = $this->reflector->reflectFunction($name);