+++ lib/WorseReflection/Core/Inference/Resolver/UnaryOpExpressionResolver.php
@@ -3,16 +3,17 @@
 namespace Phpactor\WorseReflection\Core\Inference\Resolver;
 
 use Microsoft\PhpParser\Node;
+use Microsoft\PhpParser\Node\Expression\BinaryExpression;
 use Microsoft\PhpParser\Node\Expression\UnaryExpression;
 use Microsoft\PhpParser\TokenKind;
 use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Inference\NodeContext;
 use Phpactor\WorseReflection\Core\Inference\NodeContextFactory;
+use Phpactor\WorseReflection\Core\Inference\NodeContextModifier;
 use Phpactor\WorseReflection\Core\Inference\NodeContextResolver;
 use Phpactor\WorseReflection\Core\Inference\Resolver;
 use Phpactor\WorseReflection\Core\Type;
 use Phpactor\WorseReflection\Core\Type\BitwiseOperable;
-use Phpactor\WorseReflection\Core\Type\MissingType;
 use Phpactor\WorseReflection\Core\Util\NodeUtil;
 use Phpactor\WorseReflection\TypeUtil;
 
@@ -21,35 +22,57 @@ class UnaryOpExpressionResolver implements Resolver
     public function resolve(NodeContextResolver $resolver, Frame $frame, Node $node): NodeContext
     {
         assert($node instanceof UnaryExpression);
+        $operand = $resolver->resolveNode($frame, $node->operand);
+
+        // see sister hack in BinaryExpressionResolver
+        // https://github.com/Microsoft/tolerant-php-parser/issues/19
+        $doubleNegate = $this->shouldDoubleNegate($node);
+
         $context = NodeContextFactory::create(
             $node->getText(),
             $node->getStartPosition(),
             $node->getEndPosition(),
             [
             ]
-        );
-        $operand = $resolver->resolveNode($frame, $node->operand);
+        )->withTypeAssertions(
+            $operand->typeAssertions()
+        )->withType($operand->type());
         $operatorKind = NodeUtil::operatorKindForUnaryExpression($node);
 
-        $type = $this->resolveType($operatorKind, $operand->type());
-        return $context->withType($type);
+        return $this->resolveType($context, $operatorKind, $operand->type(), $doubleNegate);
     }
 
-    private function resolveType(int $operatorKind, Type $type): Type
+    private function resolveType(NodeContext $context, int $operatorKind, Type $type, bool $doubleNegate): NodeContext
     {
         switch ($operatorKind) {
             case TokenKind::ExclamationToken:
-                return TypeUtil::toBool($type)->negate();
+                return $doubleNegate ? $context : NodeContextModifier::negate($context);
             case TokenKind::PlusToken:
-                return TypeUtil::toNumber($type)->identity();
+                return $context->withType(TypeUtil::toNumber($type)->identity());
             case TokenKind::MinusToken:
-                return TypeUtil::toNumber($type)->negative();
+                return $context->withType(TypeUtil::toNumber($type)->negative());
             case TokenKind::TildeToken:
                 if ($type instanceof BitwiseOperable) {
-                    return $type->bitwiseNot();
+                    return $context->withType($type->bitwiseNot());
                 }
         }
 
-        return new MissingType();
+        return $context;
+    }
+
+    private function shouldDoubleNegate(UnaryExpression $node): bool
+    {
+        /** @phpstan-ignore-next-line TPTodo */
+        if (!$node->operand instanceof BinaryExpression) {
+            return false;
+        }
+
+        /** @phpstan-ignore-next-line TPTodo */
+        if (!$node->operand->leftOperand instanceof UnaryExpression) {
+            return false;
+        }
+
+        $operatorKind = NodeUtil::operatorKindForUnaryExpression($node->operand->leftOperand);
+        return $operatorKind === TokenKind::ExclamationToken;
     }
 }